cp performance chart: move options to separate files
