+++ app/javascript/components/charts/cp-performance/CPPerformance.js
@@ -1,7 +1,6 @@
 /* eslint-disable react/no-this-in-sfc */
 
 import React, { useEffect, useRef, useMemo, useState } from 'react';
-import { renderToString } from 'react-dom/server';
 import PropTypes from 'prop-types';
 
 import Highcharts from 'highcharts';
@@ -9,12 +8,9 @@ import HighchartsReact from 'highcharts-react-official';
 
 import PlusIcon from 'images/icons/plus.svg';
 
-import { groupAllAreaSeries } from './helpers';
+import { getOptions, COLORS } from './options';
 import CompanySelector from './CompanySelector';
 import CompanyTag from './CompanyTag';
-import Tooltip from './Tooltip';
-
-const COLORS = ['#00C170', '#ED3D4A', '#FFDD49', '#440388', '#FF9600', '#B75038', '#00A8FF', '#F78FB3', '#191919', '#F602B4'];
 
 // TODO: move to common hooks
 const useCallbackOutsideClick = (element, action) => {
@@ -64,18 +60,6 @@ function CPPerformance({ dataUrl, companySelector, unit }) {
   }, []);
   useCallbackOutsideClick(companySelectorWrapper, () => setCompanySelectorVisible(false));
 
-  const handleLegendItemRemove = (item) => {
-    setLegendItems(
-      applyColorsToLegendItems(
-        legendItems.filter(l => l.name !== item.name)
-      )
-    );
-  };
-
-  const handleSelectedCompaniesChange = (selected) => {
-    setLegendItems(getLegendItems(data.filter((d) => selected.includes(d.name))));
-  };
-
   const companies = useMemo(
     () => data.filter(d => d.type !== 'area').map(i => i.name).sort(),
     [data]
@@ -86,96 +70,23 @@ function CPPerformance({ dataUrl, companySelector, unit }) {
     [data, selectedCompanies]
   );
 
-  // TODO: move to separate file
-  const options = {
-    chart: {
-      height: '500px',
-      marginTop: 30,
-      events: {
-        render() {
-          groupAllAreaSeries();
-        }
-      }
-    },
-    credits: {
-      enabled: false
-    },
-    colors: COLORS,
-    legend: {
-      enabled: false
-    },
-    plotOptions: {
-      area: {
-        marker: {
-          enabled: false
-        }
-      },
-      line: {
-        marker: {
-          enabled: false
-        }
-      },
-      column: {
-        stacking: 'normal'
-      },
-      series: {
-        lineWidth: 4,
-        marker: {
-          states: {
-            hover: {
-              enabled: false
-            }
-          }
-        }
-      }
-    },
-    yAxis: {
-      title: {
-        text: unit,
-        reserveSpace: false,
-        textAlign: 'left',
-        align: 'high',
-        rotation: 0,
-        x: 0,
-        y: -20
-      }
-    },
-    xAxis: {
-      crosshair: {
-        width: 2,
-        color: '#191919'
-      }
-    },
-    title: {
-      text: ''
-    },
-    tooltip: {
-      crosshairs: true,
-      shared: true,
-      useHTML: true,
-      formatter() {
-        const xValue = this.x;
-        const yValues = this.points.map(p => ({
-          value: p.y,
-          color: p.color,
-          title: p.series.name,
-          dashStyle: p.point.zone.dashStyle,
-          isTargeted: p.point.zone.dashStyle === 'dot',
-          isBenchmark: p.series.type === 'area'
-        }));
-
-        return renderToString(
-          <Tooltip xValue={xValue} yValues={yValues} unit={unit} />
-        );
-      }
-    },
-    series: chartData
-  };
-
+  // handlers
   const handleAddCompaniesClick = (e) => {
     if (e.currentTarget) e.currentTarget.blur();
     setCompanySelectorVisible(!showCompanySelector);
   };
+  const handleLegendItemRemove = (item) => {
+    setLegendItems(
+      applyColorsToLegendItems(
+        legendItems.filter(l => l.name !== item.name)
+      )
+    );
+  };
+  const handleSelectedCompaniesChange = (selected) => {
+    setLegendItems(getLegendItems(data.filter((d) => selected.includes(d.name))));
+  };
+
+  const options = getOptions({ chartData, unit });
 
   return (
     <div className="chart chart--cp-performance">
